Move to citric.rest
