+++ src/citric/rest/__init__.py
@@ -1,3 +1,5 @@
+"""LimeSurvey REST API client."""
+
 from __future__ import annotations
 
 from .client import RESTClient
